Relocate net_capital_gains variable
